pyproject.toml
[project]
name = "dieseldusel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.149.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "groq>=0.11.0",
    "langchain-groq>=0.2.1",
    "langchain>=0.3.7",
    "openai>=1.53.1",
    "pillow>=11.0.0",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=21.6",
    "pytz>=2024.2",
]
